refactor(boxws): split boxws.go into client/frames/teardown/thumb/dispatch
Move-only step 1 of catalog #23; handleMessage moved as one untouched
block so the scan-before-switch ordering is provably unchanged.
